+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Next
+
+* [ENHANCEMENT] Remove the `react-transition-group` dependency. (#56)
+
 ### v3.0.2 (7 October 2017)
 
 * [BUGFIX] Fix updating children when the current child re-renders itself. (#53)